requirements.test.txt
pytest-homeassistant-custom-component
pytest-cov
pytest-mock
